test: make sure that TmpTigerBeetle prints stderr in case of an error

+++ src/testing/tmp_tigerbeetle.zig
@@ -19,6 +19,12 @@ port_str: stdx.BoundedArray(u8, 8),
 
 tmp_dir: std.testing.TmpDir,
 
+// A separate thread for reading process stderr without blocking it. The process must be terminated
+// before stopping the StreamReader.
+//
+// StreamReader echoes process' stderr on exit unless explicitly instructed otherwise.
+stderr_reader: *StreamReader,
+
 process: std.process.Child,
 
 pub fn init(
@@ -68,19 +74,26 @@ pub fn init(
         .{ .tigerbeetle = tigerbeetle, .data_file = data_file },
     );
 
+    var reader_maybe: ?*StreamReader = null;
     // Pass `--addresses=0` to let the OS pick a port for us.
     var process = try shell.spawn(
         .{
             .stdin_behavior = .Pipe,
             .stdout_behavior = .Pipe,
-            .stderr_behavior = .Ignore,
+            .stderr_behavior = .Pipe,
         },
         "{tigerbeetle} start --development --addresses=0 {data_file}",
         .{ .tigerbeetle = tigerbeetle, .data_file = data_file },
     );
     errdefer {
-        _ = process.kill() catch unreachable;
+        if (reader_maybe) |reader| {
+            reader.stop(gpa, &process); // Will log stderr.
+        } else {
+            _ = process.kill() catch unreachable;
+        }
     }
+
+    reader_maybe = try StreamReader.start(gpa, process.stderr.?);
 
     const port = port: {
         var exit_status: ?std.process.Child.Term = null;
@@ -106,12 +119,78 @@ pub fn init(
         .port = port,
         .port_str = port_str,
         .tmp_dir = tmp_dir,
+        .stderr_reader = reader_maybe.?,
         .process = process,
     };
 }
 
 pub fn deinit(tb: *TmpTigerBeetle, gpa: std.mem.Allocator) void {
-    _ = gpa;
-    _ = tb.process.kill() catch unreachable;
+    if (tb.stderr_reader.log_stderr.load(.seq_cst) == .on_early_exit) {
+        tb.stderr_reader.log_stderr.store(.no, .seq_cst);
+    }
+    assert(tb.process.term == null);
+    tb.stderr_reader.stop(gpa, &tb.process);
+    assert(tb.process.term != null);
     tb.tmp_dir.cleanup();
 }
+
+pub fn log_stderr(tb: *TmpTigerBeetle) void {
+    tb.stderr_reader.log_stderr.store(.yes, .seq_cst);
+}
+
+const StreamReader = struct {
+    const LogStderr = std.atomic.Value(enum(u8) { no, yes, on_early_exit });
+
+    log_stderr: LogStderr = LogStderr.init(.on_early_exit),
+    thread: std.Thread,
+    file: std.fs.File,
+
+    pub fn start(gpa: std.mem.Allocator, file: std.fs.File) !*StreamReader {
+        var result = try gpa.create(StreamReader);
+        errdefer gpa.destroy(result);
+
+        result.* = .{
+            .thread = undefined,
+            .file = file,
+        };
+
+        result.thread = try std.Thread.spawn(.{}, thread_main, .{result});
+        return result;
+    }
+
+    pub fn stop(self: *StreamReader, gpa: std.mem.Allocator, process: *std.process.Child) void {
+        // Shutdown sequence is tricky:
+        // 1. Terminate the process, but _don't_ close our side of the pipe.
+        // 2. Wait until the thread exits.
+        // 3. Close stderr file descriptor.
+        // TODO(Zig) https://github.com/ziglang/zig/issues/16820
+        if (builtin.os.tag == .windows) {
+            const exit_code = 1;
+            std.os.windows.TerminateProcess(process.id, exit_code) catch {};
+        } else {
+            std.posix.kill(process.id, std.posix.SIG.TERM) catch {};
+        }
+        assert(process.stderr != null);
+        self.thread.join();
+        _ = process.wait() catch unreachable;
+        assert(process.stderr == null);
+        gpa.destroy(self);
+    }
+
+    fn thread_main(reader: *StreamReader) void {
+        // NB: Zig allocators are not thread safe, so use mmap directly to hold process' stderr.
+        const allocator = std.heap.page_allocator;
+
+        var buffer = std.ArrayList(u8).init(allocator);
+        defer buffer.deinit();
+
+        // NB: don't use `readAllAlloc` to get partial output in case of errors.
+        reader.file.reader().readAllArrayList(&buffer, 100 * 1024 * 1024) catch {};
+        switch (reader.log_stderr.load(.seq_cst)) {
+            .on_early_exit, .yes => {
+                log.err("tigerbeetle stderr:\n++++\n{s}\n++++", .{buffer.items});
+            },
+            .no => {},
+        }
+    }
+};
